Fix busy loop when there is audio track
When there is audio track, the stream buffer may not be empty and it can't send frame if there is only on track.
In this case, it would be in a busy loop which cause CPU usage high.
It can be fix by adding a send count.  If it doesn't send anything, then it can sleep a little.